feat: run test before deploy

+++ deploy.sh
@@ -1,3 +1,11 @@
 #! /bin/bash
 
-git push --follow-tags --force
+# Run flutter test
+flutter test
+
+# Check if the command was successful
+if [ $? -eq 0 ]; then
+    # Push changes to GitHub 
+    git push --follow-tags --force
+
+fi
